requirements.txt
Flask==3.0.3
google-api-python-client==2.129.0
google-auth-oauthlib==1.2.0
groq==0.5.0
pre-commit==3.7.1
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
wandb==0.17.0
youtube-transcript-api==0.6.2